Zlata fix some moment add libs
