--- client/index.html
@@ -1,30 +1,0 @@
-<!doctype html>
-<html>
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <title>JS/Socket.IO chat</title>
-
-    <link rel="stylesheet" href="/style/font-awesome-4.7.0/css/font-awesome.min.css">
-    <link rel="stylesheet" type="text/css" href="/style/normalize.css"/>
-    <link rel="stylesheet" type="text/css" href="/style/grid.css">
-    <link rel="stylesheet" type="text/css" href="/style/main.css">
-
-    <link href="https://afeld.github.io/emoji-css/emoji.css" rel="stylesheet">
-    <link href="https://fonts.googleapis.com/css?family=Lato:400,700&amp;subset=latin-ext" rel="stylesheet">
-    <link href="https://fonts.googleapis.com/css?family=Shrikhand" rel="stylesheet">
-</head>
-<body>
-<div id="root">
-
-
-</div>
-
-<script src="/index.js"></script>
-
-<script src="https://cdn.socket.io/socket.io-1.2.0.js"></script>
-<script>
-    var socket = io();
-</script>
-</body>
-</html>
