+++ style/font-awesome-4.7.0/less/stacked.less
