+++ style/font-awesome-4.7.0/scss/_icons.scss
